add qq version 8.8.17

+++ com.tencent.mobileqq/classes.jar/com/tencent/mobileqq/mini/utils/ReportLogUtil$1.java
@@ -11,10 +11,13 @@ final class ReportLogUtil$1
   
   public void onFail(String paramString)
   {
-    paramString = new StringBuilder();
-    paramString.append("upload log fail: ");
-    paramString.append(this.val$logPath);
-    QLog.d("ReportLogUtil", 1, paramString.toString());
+    if (QLog.isColorLevel())
+    {
+      paramString = new StringBuilder();
+      paramString.append("upload log fail: ");
+      paramString.append(this.val$logPath);
+      QLog.d("ReportLogUtil", 2, paramString.toString());
+    }
     paramString = new File(this.val$logPath);
     if (paramString.exists()) {
       paramString.delete();
@@ -23,10 +26,13 @@ final class ReportLogUtil$1
   
   public void onSuccess(String paramString)
   {
-    StringBuilder localStringBuilder = new StringBuilder();
-    localStringBuilder.append(" upload log success: ");
-    localStringBuilder.append(this.val$logPath);
-    QLog.d("ReportLogUtil", 1, localStringBuilder.toString());
+    if (QLog.isColorLevel())
+    {
+      StringBuilder localStringBuilder = new StringBuilder();
+      localStringBuilder.append(" upload log success: ");
+      localStringBuilder.append(this.val$logPath);
+      QLog.d("ReportLogUtil", 2, localStringBuilder.toString());
+    }
     MiniAppCmdUtil.getInstance().reportLogFileUrl(null, this.val$appid, paramString, this.val$fileSize, new ReportLogUtil.1.1(this, paramString));
     paramString = new File(this.val$logPath);
     if (paramString.exists()) {
@@ -36,7 +42,7 @@ final class ReportLogUtil$1
 }
 
 
-/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes15.jar
+/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes22.jar
  * Qualified Name:     com.tencent.mobileqq.mini.utils.ReportLogUtil.1
  * JD-Core Version:    0.7.0.1
  */
